Skeleton for assignements
