--- controllers/assigments.js
@@ -1,5 +1,0 @@
-exports.getAssigment = async (req, res) => {
-    res.status(200).json({
-        message: 'Assigment page is working',
-    })
-};
