+++ routes/assignments.js
@@ -1,0 +1,19 @@
+const express = require('express');
+const router = express.Router();
+
+
+//import controllers
+const { getAssigment, getSpecificAssigment, getNumberOfQuestions } = require('../controllers/assignments');
+
+//import middlewares
+const isAuth = require('../middleware/is-auth');
+
+
+// api routes
+router.get('/', isAuth, getAssigment);
+
+router.get('/specificAssignement', isAuth, getSpecificAssigment);
+
+router.get('/numberOfQuestions', isAuth, getNumberOfQuestions);
+
+module.exports = router;
